Fix Duplicate Class Warning
Gluten uses the current context class loader to list all the resources
from the classpath:
  Thread.currentThread.getContextClassLoader.resources(...)

But Spark's classloader is a hierarchy:
Spark-shell's classloader
        |
        |
       \ /
        .
 Driver Classloader

Due to this hierarchy, every class appears twice, hence we get the
duplicate classes warning.

This is not really a issue, but to fix this, I am only taking distinct
classes. If there was conflict as below

1./path/to/jar1.jar!/META-INF/gluten-components
2./path/to/jar1.jar!/META-INF/gluten-components

This would return only jar since there's only one unique jar

But if there was gluten-components/ in two different jars, this would
still return a warning since they are two different jars

Change-Id: I42f43a6cd8d87a20abb219bd1548effde66340b3
Reviewed-on: https://bigdataoss-internal-review.googlesource.com/c/third_party/apache/incubator-gluten/+/116875
Tested-by: Prow Service Account <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Preetesh Verma <preeteshverma@google.com>
Reviewed-by: Abhinav Koul <abhinavkoul@google.com>