Activate BQ component only when Spark-BQ connector is present
Velox Bigquery component fails with ClassNotFoundException when BQ jar
is not present in classpath. This can happen for multiple reasons, with
the primary reason being the user doesn't intend to do a bigquery read.
But we enable the component anyways since we cannot activate it at
runtime, it's a build time setting. Instead, this change injects the
rule only if the class is present.

Change-Id: I59332beacbce2b14d2bfcc287b0910070bbf8307
Reviewed-on: https://bigdataoss-internal-review.googlesource.com/c/third_party/apache/incubator-gluten/+/127835
Tested-by: Prow Service Account <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Adesh Rao <adeshr@google.com>
Reviewed-by: Revanth Venkat Mikkilineni <revanthvenkat@google.com>